requirements.txt
Flask
flask-cors
mariadb
mysql-connector